src/VumbaSoft.AdventureWorks.Controllers/Sales/SalesTerritories/SalesTerritoriesController.cs
using Microsoft.AspNetCore.Mvc;
using VumbaSoft.AdventureWorks.Objects;
using VumbaSoft.AdventureWorks.Services;
using VumbaSoft.AdventureWorks.Validators;
using System;

namespace VumbaSoft.AdventureWorks.Controllers.Sales
{
    [Area("Sales")]
    public class SalesTerritoriesController : ValidatedController<ISalesTerritoryValidator, ISalesTerritoryService>
    {
        public SalesTerritoriesController(ISalesTerritoryValidator validator, ISalesTerritoryService service)
            : base(validator, service)
        {
        }

        [HttpGet]
        public ViewResult Index()
        {
            return View(Service.GetViews());
        }

        [HttpGet]
        public ViewResult Create()
        {
            return View();
        }

        [HttpPost]
        public ActionResult Create(SalesTerritoryView territory)
        {
            if (!Validator.CanCreate(territory))
                return View(territory);

            Service.Create(territory);

            return RedirectToAction("Index");
        }

        [HttpGet]
        public ActionResult Details(Int32 id)
        {
            return NotEmptyView(Service.Get<SalesTerritoryView>(id));
        }

        [HttpGet]
        public ActionResult Edit(Int32 id)
        {
            return NotEmptyView(Service.Get<SalesTerritoryView>(id));
        }

        [HttpPost]
        public ActionResult Edit(SalesTerritoryView territory)
        {
            if (!Validator.CanEdit(territory))
                return View(territory);

            Service.Edit(territory);

            return RedirectToAction("Index");
        }

        [HttpGet]
        public ActionResult Delete(Int32 id)
        {
            return NotEmptyView(Service.Get<SalesTerritoryView>(id));
        }

        [HttpPost]
        [ActionName("Delete")]
        public RedirectToActionResult DeleteConfirmed(Int32 id)
        {
            Service.Delete(id);

            return RedirectToAction("Index");
        }
    }
}
